requirements.txt
numpy
matplotlib
pandas
requests
sklearn
flask
django
opencv-python
pypcapfile
termcolor
nltk
pyyaml